+++ server/index.js
@@ -11,12 +11,12 @@ dotenv.config();
 connectDB();
 
 var corsoption = {
-    origin: 'https://bp-gc.in'
-}
+  origin: 'https://bp-gc.in',
+};
 // Middleware functions
 app.use(cors(corsoption));
 app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended:true }));
+app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(express.static(__dirname));
 // Define Routes
@@ -25,6 +25,6 @@ app.use('/api/url', require('./routes/url'));
 app.use('/api/user', require('./routes/user'));
 app.use('/api/login/google', require('./routes/googleAuth'));
 
-const PORT = 5000;
+const PORT = 5010;
 
 app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
